Split VB files; use Combine Files VScode extension
Although the code for each report ultimately needs to be in one file, it
becomes impossible to work on the code this way for a couple reasons:
* Code that needs to be shared between reports can get out of sync
existing in multiple places
* It's hard to work in a file >100 lines

This commit introduces fake modularity by splitting the code into
multiple files according to function, with more generally-applicable
code being put in the `utilities` directory.

Contributers will probably want to add the Combine Files VScode
extension to work on this project in the future. Obviously this
significantly ties the project to VScode / Visual Studio, but since
we're using .Net anyway, using alternative editors seemed unlikely.

`.combinefilesrc.json` contains the config settings for the Combine
Files extension. Briefly:
* All source files in the categories named for the reports are included
in the output. To include a new utility, add a glob for it.
* To produce the outputs, right-click the parent folder or any blank
space in VScode's Explorer window, and click `Combine Files`.

NOTE: There appears to be a bug in the extension causing it to only
combine the last file-group in the fileGroups array. I will look into
fixing this and submitting a pull request to the project.