Slow but working lookup feature for mark12 grades
This implementation is designed to be used with a grades parameter that
maps group|learner#column => grade. Grades for each student are then
looked up and added to the store, while returning their count. Then, in
the body of the report, Grades are lookup up from the store.

This implementation proved too slow, and I suspect that there won't be
possible to show all students' grades in one report with reasonable
rendering time.

I'll try using drill-down reports next.

This also includes functions that enables the assessment-config
lookup-table to pass information to the the main dataset so that it
knows which groups are only mark6. This will likely stay.