Escape regex meta-chars in CountMatchingParams
`|` is a commonly used delimiter in param values that are concatenations of
fields. The regex engine interprets this as `OR`. And worse, if the `|` is at
the end of the search-string, all test strings returned true.

The new `EscapeRegexString` is now used to escape all single-character regex
meta-characters in user-supplied strings so that a regex can be built from the
string that matches the supplied characters literally.