Find all problems with Report 6 ranks & grades
Redesigned the way the custom code is called from the report, as well as the
interface it presents. Now, all the analysis is done at the dataset level, with
the results being stored as fields in the main dataset. As a result, data is not
lost when the report body decides to rerender (when this happens, you lose any
VB objects, which is why you can't rely on accessing them from the report body).

The main DS uses the VB to populate 2 fields:
* ProblemCode: an integer whose bits represent these booleans:
    - 1st bit => has multiple, conflicting ranks
    - 2nd bit => has same ranks as another learner
    - 3rd bit => has skipped a rank
* RankDelta: The minimum magnitude (signed) change in rank needed for the
    learner to have a rank appropriate for their grade. Note that this may place
    them in the same rank as another learner

GroupID is used in the method calls, instead of GroupCode, primarily because
codes wouldn't be used by the dataset for any other reason, while IDs would.