Enable match-strategies in param functions
Add match-strategies options to all param-lookup functions (for which
doing so makes sense).

Eschew Optional params in favour of Overloads

Breaking change for reports using the old syntax of passing 0, 1 or 2 to
determine the match strategy for the CountMatchingParams function.

Moved the helper functions:
* StartsWithRegex
* EscapeRegexString
* ThrowUnlessSearchesAreSearchable
to an independent file, param_helpers.vb, since they're shared by
functions in lookup_params.vb and search_params.vb. The former must be
combined if either of the latter are used in a report.

Also added shortcut specifically for Report-6, which looks up all of a
learner's marks for a column, and joins all those that differ. It has an
overload so that the search item can passed in 2 parts; if the second
part is blank, the param isn't searched.