Lookup Params using Regular Expressions
Add new `matchStrategy`, 'R', to LookupParams function, which uses the
`searchItem` to compile a regex to search through the param. Since the
pattern is so similar to strategy 'S' ('starts-with'), I extracted a
method, `SearchUsingRegex` to do the work.

I decoupled the two match-strategy type-checks: one on the `searchItem`
(which must be String) and one on the searches array (which must be
String()). I split these into different methods which each call another
method to produce an exception message.

The two new type checks can still be combined using one method, but I
changed its signature, adding a `matchStrategy` param, to make error
messages clearer.

After discovering that a postfix `C` is used to create a Char literal
(e.g. "R"C), I updated all method calls where a matchStrategy is
specified, to ensure Char is passed, not a String. With a String the
compiler will warn that it could be cast to an Integer (which would use
another overload for a lot of the methods in this library), even though
this would presumably not happen in practice.

I also altered method calls where "E" is passed unnecessarily as a
`matchStrategy`, since Equals is the default strategy anyway.