Fix M12 science/double-science bug
Except for GCSE Triple Science and A Level Sciences, learners ARE in
multiple groups for science, violating one of the assumptions behind the
SLT & HoD M12 reports. This caused all the grades to be attributed to
the first group, (and none to subsequent groups).

I fixed this using a custom function that checks whether the subject is
science, and if so, looks among the matching leaner-subjects in the
lookup for a group with the same denomination (e.g. Bi/Ch/Ph).